+++ src/classes.js
@@ -1,0 +1,56 @@
+'use strict';
+
+var trim = /^\s+|\s+$/g;
+var whitespace = /\s+/g;
+
+function interpret (input) {
+  return typeof input === 'string' ? input.replace(trim, '').split(whitespace) : input;
+}
+
+function classes (node) {
+  return node.className.replace(trim, '').split(whitespace);
+}
+
+function set (node, input) {
+  node.className = input.join(' ');
+}
+
+function add (node, input) {
+  var current = remove(node, input);
+  var values = interpret(input);
+
+  current.push.apply(current, values);
+  set(node, values);
+}
+
+function remove (node, input) {
+  var current = classes(node);
+  var values = interpret(input);
+
+  values.forEach(function (value) {
+    var i = current.indexOf(value);
+    if (i !== -1) {
+      current.splice(i, 1);
+      set(node, current);
+    }
+  });
+
+  return current;
+}
+
+function contains (node, input) {
+  var current = classes(node);
+  var values = interpret(input);
+
+  return values.every(function (value) {
+    return current.indexOf(value) !== -1;
+  });
+}
+
+module.exports = {
+  add: add,
+  remove: remove,
+  contains: contains,
+  set: set,
+  get: classes
+};