+++ src/events.js
@@ -17,10 +17,10 @@ function addEventEasy (element, evt, fn) {
 
 function addEventHard (element, evt, fn) {
   return element.attachEvent('on' + evt, function (e) {
-    var e = e || window.event;
+    e = e || window.event;
     e.target = e.target || e.srcElement;
-    e.preventDefault  = e.preventDefault  || function preventDefault () { e.returnValue = false; }
-    e.stopPropagation = e.stopPropagation || function stopPropagation () { e.cancelBubble = true; }
+    e.preventDefault  = e.preventDefault  || function preventDefault () { e.returnValue = false; };
+    e.stopPropagation = e.stopPropagation || function stopPropagation () { e.cancelBubble = true; };
     fn.call(element, e);
   });
 }