+++ src/public.js
@@ -2,10 +2,20 @@
 
 var dom = require('./dom');
 var core = require('./core');
+var Dominus = require('./Dominus.ctor');
+var tag = /^\s*<([a-z]+(?:-[a-z]+)?)\s*\/?>\s*$/i;
 
 function api (selector, context) {
-  if (arguments.length < 2 && typeof selector !== 'string') {
+  var notText = typeof selector !== 'string';
+  if (notText && arguments.length < 2) {
     return core.cast(selector);
+  }
+  if (notText) {
+    return new Dominus();
+  }
+  var matches = selector.match(tag);
+  if (matches) {
+    return dom.make(matches[1]);
   }
   return api.find(selector, context);
 }